Add JSON Schema and code to parse it
Added the new JSON schema as a first pass at the expected data storage format.  It was designed around keeping the individual analysis pieces front and center, and dangle the other information off of it.

The added makefile turns the schema into a Go parser file, with some necessary cleanup.

There's also the beginnings of a CLI tool.